refactor: Convert Attribute class to domain class (instead of dto class)
* Attribute is a very prominent and often used class
* We don't want to duplicate this class all over the place
* Serialization is done via a converter anyways
* Saving the conversion is such a performance gain that it makes sense to realize that gain
