+++ src/Api.Rest.Dtos.Tests/Data/DataValueDtoTest.cs
@@ -13,6 +13,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 	#region usings
 
 	using NUnit.Framework;
+	using Zeiss.PiWeb.Api.Contracts;
 	using Zeiss.PiWeb.Api.Rest.Dtos.Data;
 
 	#endregion
@@ -48,8 +49,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_Compare_Values_When_Attributes_Count_Is_Different()
 		{
 			// arrange
-			var value1 = new DataValueDto( new[] { new AttributeDto( 42, "foo" ), new AttributeDto( 24, "bar" ) } );
-			var value2 = new DataValueDto( new[] { new AttributeDto( 24, "foo" ) } );
+			var value1 = new DataValueDto( new[] { new Attribute( 42, "foo" ), new Attribute( 24, "bar" ) } );
+			var value2 = new DataValueDto( new[] { new Attribute( 24, "foo" ) } );
 
 			// act, assert
 			Assert.That( Equals( value1, value2 ), Is.False );
@@ -81,8 +82,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_Compare_Values_When_Attributes_Are_Equal()
 		{
 			// arrange
-			var value1 = new DataValueDto( new[] { new AttributeDto( 1, 5 ), new AttributeDto( 2, "foo" ) } );
-			var value2 = new DataValueDto( new[] { new AttributeDto( 1, 5 ), new AttributeDto( 2, "foo" ) } );
+			var value1 = new DataValueDto( new[] { new Attribute( 1, 5 ), new Attribute( 2, "foo" ) } );
+			var value2 = new DataValueDto( new[] { new Attribute( 1, 5 ), new Attribute( 2, "foo" ) } );
 
 			// act, assert
 			Assert.That( Equals( value1, value2 ), Is.True );
@@ -92,8 +93,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_Compare_Values_When_Attributes_Are_Not_Equal()
 		{
 			// arrange
-			var value1 = new DataValueDto( new[] { new AttributeDto( 1, 5 ), new AttributeDto( 2, "foo" ) } );
-			var value2 = new DataValueDto( new[] { new AttributeDto( 1, 4 ), new AttributeDto( 2, "bar" ) } );
+			var value1 = new DataValueDto( new[] { new Attribute( 1, 5 ), new Attribute( 2, "foo" ) } );
+			var value2 = new DataValueDto( new[] { new Attribute( 1, 4 ), new Attribute( 2, "bar" ) } );
 
 			// act, assert
 			Assert.That( Equals( value1, value2 ), Is.False );
@@ -146,8 +147,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		[Test]
 		public void Test_GetHashCode_IsDifferent_When_Attributes_Count_Are_Different()
 		{
-			var value1 = new DataValueDto( new[] { new AttributeDto( 42, "foo" ), new AttributeDto( 24, "bar" ) } );
-			var value2 = new DataValueDto( new[] { new AttributeDto( 24, "foo" ) } );
+			var value1 = new DataValueDto( new[] { new Attribute( 42, "foo" ), new Attribute( 24, "bar" ) } );
+			var value2 = new DataValueDto( new[] { new Attribute( 24, "foo" ) } );
 
 			// act
 			var hashCode1 = value1.GetHashCode();
@@ -161,8 +162,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_GetHashCode_IsSame_When_Attributes_Are_Equal()
 		{
 			// arrange
-			var value1 = new DataValueDto( new[] { new AttributeDto( 1, 5 ), new AttributeDto( 2, "foo" ) } );
-			var value2 = new DataValueDto( new[] { new AttributeDto( 1, 5 ), new AttributeDto( 2, "foo" ) } );
+			var value1 = new DataValueDto( new[] { new Attribute( 1, 5 ), new Attribute( 2, "foo" ) } );
+			var value2 = new DataValueDto( new[] { new Attribute( 1, 5 ), new Attribute( 2, "foo" ) } );
 
 			// act
 			var hashCode1 = value1.GetHashCode();
@@ -176,8 +177,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_GetHashCode_IsDifferent_When_Attributes_Are_Not_Equal()
 		{
 			// arrange
-			var value1 = new DataValueDto( new[] { new AttributeDto( 1, 5 ), new AttributeDto( 2, "foo" ) } );
-			var value2 = new DataValueDto( new[] { new AttributeDto( 1, 4 ), new AttributeDto( 2, "bar" ) } );
+			var value1 = new DataValueDto( new[] { new Attribute( 1, 5 ), new Attribute( 2, "foo" ) } );
+			var value2 = new DataValueDto( new[] { new Attribute( 1, 4 ), new Attribute( 2, "bar" ) } );
 
 			// act
 			var hashCode1 = value1.GetHashCode();
